+++ util/stack/from_args.c
@@ -6,7 +6,7 @@
 /*   By: zleullie <zleullie@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/05 16:36:05 by zleullie          #+#    #+#             */
-/*   Updated: 2026/06/06 16:48:57 by zleullie         ###   ########.fr       */
+/*   Updated: 2026/06/16 17:46:04 by zleullie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 int	stack_from_list(struct s_args *args, t_stack *a)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < args->size)
